ml/plant_disease_classifier/requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
transformers
huggingface-hub
